rename getLanguageCards and create validator function for one-pass aggregation
